\docType{data}
\name{beta}
\alias{beta}
\title{A beta value matrix}
\usage{
data(beta)
}
\description{
A beta value matrix for the GSE149282 dataset.
}
\details{
 \itemize{
      \item A matrix with 177057 rows and 24 columns.The colnames of the matrix are samples name with GSM number,the rownames of the matrix are cg sites.But it only has 177057 cg sites.
 }
}
\keyword{datasets}
