\docType{data}
\name{cpgTogene_message}
\alias{cpgTogene_message}
\title{Three dataframes of genes on the methylation microarray platform correspond to CpG sites in their promoter regions.}
\usage{
data(cpgTogene_message)
}
\description{
This contains three dataframes of genes on the methylation microarray platform correspond to CpG sites in their promoter regions.They are 'cpg.450','cpg.850','cpg.integrated'.
}
\details{
 \itemize{
      \item cpg.450. A matrix with 148076 rows and 2 columns.A matrix of genes on the 450K platform correspond to CpG sites in their promoter regions.The first column is the CpG sites and the second column is the gene symbols.
      \item cpg.850. A matrix with 217130 rows and 2 columns.A matrix of genes on the 850K platform correspond to CpG sites in their promoter regions.The first column is the CpG sites and the second column is the gene symbols.
      \item cpg.integrated. A matrix with 119306 rows and 2 columns.A matrix of genes on the integrated platform correspond to CpG sites in their promoter regions.The first column is the CpG sites and the second column is the gene symbols.
      }
   }
\keyword{datasets}
