\docType{data}
\name{example}
\alias{example}
\title{Several data sets are used to demonstrate the code.}
\usage{
data(example)
}
\description{
It contains several data sets, which uesd to demonstrate the code.These data sets are 'clininf','data1','data2' and 'normal'.
}
\details{
 \itemize{
      \item clininf. A dataframe with 24 rows and 2 columns.A dataframe with sample clinical information for the GSE149282 dataset.
      \item data1. A dataframe with 23044 rows and 97 columns.A dataframe for the function \code{Selrefergene}.It cantains 48 normal control samples and 48 colon cancer samples.The first column is the gene symbols and the remaining columns are the gene methylation values of the samples.It comes from the 850K platform.
      \item data2. A dataframe with 23037 rows and 65 columns.Another dataframe for the function \code{Selrefergene}.It cantains 32 normal control samples and 32 colon cancer samples.The first column is the gene symbols and the remaining columns are the gene methylation values of the samples.It comes from the 850K platform.
      \item normal. A dataframe with 23035 rows and 87 columns.A dataframe for the function \code{RankDMG}.It cantains 86 normal control samples.The first columns is the gene symbols and the remaining columns are the gene methylation values of the samples.It comes from the 850K platform.
 }
}
\keyword{datasets}
